Fetch whole body rather than just 1024 bytes for excerpt
This means we'll download extra information (once, as is cached) but this
will fix parsing of RTL languages where trimming the content LTR results
in bad unicode points.

In addition to this there is likely a speedup / reduction in server
load as requesting the whole body is more common. With proper caching
of summaries the extra bandwidth should only be a one time thing.